requirements.txt
telethon==1.37.0
google-generativeai==0.8.4
python-dotenv==1.0.1
